requirements.txt
asgiref==3.2.10
Django==3.1.2
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-tinymce==3.4.0
djangorestframework==3.12.4
jsonfield==3.1.0
mod-wsgi==4.9.0
Pillow==8.0.1
psycopg2==2.9.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
swapper==1.1.2.post1
whitenoise==5.3.0

